restructured script panel to comply with the MVC pattern
